server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
langsmith>=0.2.0
